getAXDRLength in DLMSUtils should return an int instead of a long value. An int allows AXDR objects with a length or 2 Gb (2 147 483 648 bytes), which i assume is more than enough us :-) Returning an int will prevent a lot of (unsafe) casts from long to int. Added JUnit tests to make sure I did not brake anything.
git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@55967 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/dlms/src/main/java/com/energyict/dlms/DLMSUtils.java
@@ -80,15 +80,15 @@ public final class DLMSUtils {
      * @param offset    The offset in the byteArray from which to start reading the length
      * @return the decimal converted length
      */
-    public static long getAXDRLength(byte[] bytes, int offset) {
-        long length = 0;
+    public static int getAXDRLength(byte[] bytes, int offset) {
+        int length = 0;
         if ((bytes[offset] & (byte) 0x80) != 0) {
             int nrOfBytes = (bytes[offset] & 0x7F);
             for (int i = 0; i < nrOfBytes; i++) {
-                length |= (((long) bytes[offset + i + 1] & 0xFF) << (8 * ((nrOfBytes - 1) - i)));
+                length |= ((bytes[offset + i + 1] & 0xFF) << (8 * ((nrOfBytes - 1) - i)));
             }
         } else {
-            length = ((long) bytes[offset] & 0xFF);
+            length = bytes[offset] & 0xFF;
         }
         return length;
     }
